% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{assumptions.dfr_dist}
\alias{assumptions.dfr_dist}
\title{Retrieve the assumptions a DFR distribution makes about the data}
\usage{
\method{assumptions}{dfr_dist}(model, ...)
}
\arguments{
\item{model}{A \code{dfr_dist} object}

\item{...}{Additional arguments (ignored)}
}
\value{
A character vector of model assumptions
}
\description{
Returns a list of assumptions that the dynamic failure rate distribution
model makes about the underlying data and process.
}
