% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{cdf.dfr_dist}
\alias{cdf.dfr_dist}
\title{Method for obtaining the cdf of a \code{dfr_dist} object.}
\usage{
\method{cdf}{dfr_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the cdf of.}

\item{...}{Additional arguments to pass into the \code{cum_haz}
constructor.}
}
\value{
A function that computes the cdf of the distribution.
It accepts \code{t}, the time at which to compute the cdf, \code{par},
the parameters of the distribution, \code{log.p} argument that
determines whether to compute the log of the cdf, \code{lower.limit},
whether to compute the lower limit (F(t)) or upper limit
(S(t) = 1-F(t)). Finally, it passes any additional arguments \code{...}
to the \code{rate} function of the \code{dfr_dist} object \code{x}.
}
\description{
Method for obtaining the cdf of a \code{dfr_dist} object.
}
