% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{cum_haz}
\alias{cum_haz}
\title{Method for obtaining the cumulative hazard function of an object.}
\usage{
cum_haz(x, ...)
}
\arguments{
\item{x}{The object to obtain the cumulative hazard function of.}

\item{...}{Additional arguments to pass.}
}
\value{
A function that computes the cumulative hazard function of the
distribution.
}
\description{
Method for obtaining the cumulative hazard function of an object.
}
