% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{density.dfr_dist}
\alias{density.dfr_dist}
\title{Method for obtaining the density (pdf) of a \code{dfr_dist} object.}
\usage{
\method{density}{dfr_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the density of.}

\item{...}{Additional arguments to pass.}
}
\value{
A function that computes the density of the distribution.
It accepts \code{t}, the time at which to compute the density, \code{par} is
the parameters of the distribution, and \code{log} determines whether to
compute the log of the density. Finally, it passes any additional
arguments \code{...} to the \code{rate} function of the \code{dfr_dist} object \code{x}.
}
\description{
Method for obtaining the density (pdf) of a \code{dfr_dist} object.
}
