% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_utils.R
\name{has_femtograd}
\alias{has_femtograd}
\title{Check if femtograd is available}
\usage{
has_femtograd()
}
\value{
TRUE if femtograd is installed
}
\description{
Check if femtograd is available
}
\keyword{internal}
