% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{hazard.dfr_dist}
\alias{hazard.dfr_dist}
\title{Method for obtaining the hazard function of a \code{dfr_dist} object.}
\usage{
\method{hazard}{dfr_dist}(x)
}
\arguments{
\item{x}{The object to obtain the hazard function of.}
}
\value{
A function that computes the hazard function of the distribution.
It accepts \code{t}, the time at which to compute the hazard function, and
\code{par}, the parameters of the distribution. If \code{par} is \code{NULL}, then
the parameters of the \code{dfr_dist} object \code{x} are used. It also
accepts a \code{log} argument that determines whether to compute the log of
the hazard function. Finally, it passes any additional arguments to the
\code{rate} function of the \code{dfr_dist} object \code{x}.
}
\description{
Method for obtaining the hazard function of a \code{dfr_dist} object.
}
