% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{inv_cdf.dfr_dist}
\alias{inv_cdf.dfr_dist}
\title{Method for obtaining the quantile (inverse cdf) of an object.}
\usage{
\method{inv_cdf}{dfr_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the inverse cdf of.}

\item{...}{Additional arguments to pass into \code{cdf} constructor.}
}
\value{
A function that computes the quantile of the distribution.
It accepts \code{p}, the probability at which to compute the quantile,
\code{par}, the parameters of the distribution, and \code{...}, any additional
arguments to pass into the constructed cdf.
}
\description{
Method for obtaining the quantile (inverse cdf) of an object.
}
