% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{is_dfr_dist}
\alias{is_dfr_dist}
\title{Function for determining whether an object is a \code{dfr_dist} object.}
\usage{
is_dfr_dist(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A logical value indicating whether \code{x} is a \code{dfr_dist}
object.
}
\description{
Function for determining whether an object is a \code{dfr_dist} object.
}
