% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_utils.R
\name{make_ad_loglik}
\alias{make_ad_loglik}
\title{Create AD-aware log-likelihood function for dfr_dist}
\usage{
make_ad_loglik(model, df)
}
\arguments{
\item{model}{A dfr_dist object}

\item{df}{Data frame with observations}
}
\value{
A function f(par) that works with femtograd value/dual objects
}
\description{
Returns a function that computes log-likelihood using femtograd objects,
enabling automatic differentiation.
}
\keyword{internal}
