% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{make_ad_loglik_analytical}
\alias{make_ad_loglik_analytical}
\title{Create AD-compatible log-likelihood using analytical cumulative hazard}
\usage{
make_ad_loglik_analytical(model, df)
}
\arguments{
\item{model}{A dfr_dist object with cum_haz_rate defined}

\item{df}{Data frame with observations}
}
\value{
A function f(par) compatible with femtograd
}
\description{
Create AD-compatible log-likelihood using analytical cumulative hazard
}
\keyword{internal}
