% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{params.dfr_dist}
\alias{params.dfr_dist}
\title{Method for obtaining the parameters of a \code{dfr_dist} object.}
\usage{
\method{params}{dfr_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the parameters of.}

\item{...}{Additional arguments (unused).}
}
\value{
The parameters of the distribution.
}
\description{
Method for obtaining the parameters of a \code{dfr_dist} object.
}
