% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{pdf.dfr_dist}
\alias{pdf.dfr_dist}
\title{Method for obtaining the pdf of a \code{dfr_dist} object.}
\usage{
\method{pdf}{dfr_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the pdf of.}

\item{...}{Additional arguments to pass.}
}
\value{
A function that computes the pdf of the distribution.
It accepts \code{y}, the value at which to compute the pdf, \code{t}, the time
at which to compute the pdf, \code{par} is the parameters of the
distribution, and \code{log} determines whether to compute the log of
the pdf. Finally, it passes any additional arguments \code{...} to
the \code{rate} function of the \code{dfr_dist} object \code{x}.
}
\description{
Method for obtaining the pdf of a \code{dfr_dist} object.
}
