% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{print.dfr_dist}
\alias{print.dfr_dist}
\title{Print method for \code{dfr_dist} objects.}
\usage{
\method{print}{dfr_dist}(x, ...)
}
\arguments{
\item{x}{The \code{dfr_dist} object to print.}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for \code{dfr_dist} objects.
}
