% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{sampler.dfr_dist}
\alias{sampler.dfr_dist}
\title{Sampling function for \code{dfr_dist} objects.}
\usage{
\method{sampler}{dfr_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the sampler of.}

\item{...}{Additional arguments to pass into the inverse CDF constructor.}
}
\value{
A function that samples from the distribution. It accepts
\code{n}, the number of samples to take, \code{par}, the parameters of the
distribution, and \code{...}, additional arguments passed to the quantile
function.
}
\description{
Uses inverse CDF sampling: generates uniform random values and
transforms them through the quantile function (inverse CDF).
}
