% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{sup.dfr_dist}
\alias{sup.dfr_dist}
\title{Method for retrieving the support of an object \code{x}.}
\usage{
\method{sup}{dfr_dist}(x)
}
\arguments{
\item{x}{The object to obtain the support of.}

\item{...}{Additional arguments to pass.}
}
\value{
A support object for \code{x}, an interval (0,Inf).
}
\description{
Method for retrieving the support of an object \code{x}.
}
