% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{surv.dfr_dist}
\alias{surv.dfr_dist}
\title{Method for obtaining the survival function of a \code{dfr_dist} object.}
\usage{
\method{surv}{dfr_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the survival function of.}

\item{...}{Additional arguments to pass into the \code{cum_haz}
constructor.}
}
\value{
A function that computes the survival function of the
distribution.
It accepts \code{t}, the time at which to compute the survival, \code{par},
the parameters of the distribution, \code{log.p} argument that
determines whether to compute the log of the survival, and
it passes any additional arguments into the \code{rate} function of
the \code{dfr_dist} object \code{x}.
}
\description{
Method for obtaining the survival function of a \code{dfr_dist} object.
}
