% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{dfr_dist}
\alias{dfr_dist}
\title{Concstructor for \code{dfr_dist} objects}
\usage{
dfr_dist(rate, par = NULL, eps = 0.01)
}
\arguments{
\item{rate}{A function that computes the hazard rate at time \code{t}.}

\item{par}{The parameters of the distribution. Defaults to \code{NULL},
which means that the parameters are unknown.}

\item{eps}{The epsilon update for numerical integration. Defaults to 0.01.}
}
\value{
A \code{dfr_dist} object.
}
\description{
We assume that the hazard rate is a function of time and any other
predictors. We also assume that intergate(rate(t), 0, Inf) = infinity
and that the support is (0, Inf).
}
