% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{params.dfr_dist}
\alias{params.dfr_dist}
\title{Method for obtaining the parameters of a \code{dfr_dist} object.}
\usage{
\method{params}{dfr_dist}(x, par = NULL)
}
\arguments{
\item{x}{The object to obtain the parameters of.}

\item{par}{The parameters to replace \code{NA} or \code{NULL} with.}
}
\value{
The parameters of the distribution.
}
\description{
checks to see if any of the parameters are \code{NA} or \code{NULL}
and if so, replaces them with the default values
in \code{x} (\code{x$par})
}
\details{
We may think of this function as a "constructor" for the parameters
of a \code{dfr_dist} object. If we don't know some parameters, we can
pass \code{NA} or \code{NULL} for them, and this function will replace them
with the default values in \code{x$par}.
}
