% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{sampler.dfr_dist}
\alias{sampler.dfr_dist}
\title{Sampling function for \code{dfr_dist} objects.}
\usage{
\method{sampler}{dfr_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the sampler of.}

\item{...}{Additional arguments to pass into the survival function}
}
\value{
A function that samples from the distribution. It accepts
\code{n}, the number of samples to take, \code{t} is the time at which to start
sampling, \code{par} are the parameters of the distribution, and \code{eps} is
the  update for numerical integration. Finally, we pass additional
arguments \code{...} into the hazard function.
}
\description{
Since S(t,par) = exp(-cum_hz(t,par)), we can sample from the
distribution by letting t = 0 (or some other positive number if
we want to condition on T > t_min), sampling from an exponential
distribution with \code{lambda = rate(t, par)}, and then rejecting
the sample if \code{runif(1) > S(t, par)}. If accepted, add that
observation to the sample, otherwise reject it, let \code{t = t + eps}
where \code{eps} is some small number, and repeat. We continue this
process until we have \code{n} observations for the sample.
}
